Added Twitter feed reader.
This loads the SpaceAPI directory and then proceeds to download all SpaceAPI information from the individual spaces. Once that is collected, the twitter names for the spaces are taken from the API information.

For each of the spaces on Twitter, a thread is started that periodically loads new tweets for these spaces (and starts with tweets performed in the last 30 minutes).